# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "description": "Zavod, o‘lcham, sirt turi bo‘yicha filtr va nom/zavod bo‘yicha qidiruv. Sahifalangan.\n\n🔒 Javobda narx YO‘Q va zaxiraning aniq soni YO‘Q — faqat `availability`: `AVAILABLE` yoki `UNAVAILABLE`.\n\nO‘chirilgan mahsulotlar va o‘chirilgan ZAVODNING mahsulotlari ro‘yxatga tushmaydi.",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Saralash maydoni. Ro‘yxatdagidan boshqasi qabul qilinmaydi.\n\nNarx bo‘yicha saralash YO‘Q — ochiq katalogda narx ko‘rinmaydi.",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "name",
                "createdAt",
                "viewCount"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "factoryId",
            "required": false,
            "in": "query",
            "description": "Zavod bo‘yicha filtr",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sizeId",
            "required": false,
            "in": "query",
            "description": "O‘lcham bo‘yicha filtr",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "surface",
            "required": false,
            "in": "query",
            "description": "Sirt turi: pol yoki devor",
            "schema": {
              "type": "string",
              "enum": [
                "POL",
                "DEVOR"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Qidiruv — mahsulot nomi yoki zavod nomi bo‘yicha (katta-kichik harf farqlanmaydi)",
            "schema": {
              "maxLength": 100,
              "example": "yongxin",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sahifalangan mahsulotlar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PaginatedProductListItemResponseDtoDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Mahsulotlar ro‘yxati va filtr",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/products/{slug}": {
      "get": {
        "description": "Kartadagi hamma narsa + tavsif va to‘liq media ro‘yxati (suratlar, 360° materiallar).\n\n🔒 Narx va zaxira aniq soni bu yerda ham yo‘q.\n\n⚠ Ko‘rishlar soni bu endpointda OSHMAYDI — buning uchun alohida `POST /products/:slug/view` bo‘ladi. Aks holda har bir bot tashrifi statistikani buzardi.",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Mahsulot URL nomi",
            "schema": {
              "example": "lyuks-keramogranit-60x60",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mahsulot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductDetailResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi, o‘chirilgan yoki zavodi o‘chirilgan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Mahsulot sahifasi",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/products/{slug}/similar": {
      "get": {
        "description": "Mahsulot kartochkasi ostidagi blok (TZ 3.1).\n\n1. Admin qo‘lda bog‘lagan mahsulotlar — doim birinchi.\n2. Mahsulotning o‘zi kam qolgan yoki tugagan bo‘lsa — qolgan joy avtomatik to‘ldiriladi: bir xil o‘lcham va sirt, omborda yetarli; bir xil rangdagilar oldinda.\n\nOmborda yo‘q va o‘chirilgan mahsulotlar chiqmaydi.\n\n🔒 Narx va zaxira aniq soni yo‘q — oddiy katalog kartasi.",
        "operationId": "ProductsController_findSimilar",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Mahsulot URL nomi",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 20,
              "default": 8,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "O‘xshash mahsulotlar (bo‘sh bo‘lishi mumkin)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ProductListItemResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi yoki o‘chirilgan",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Shunga o‘xshash mahsulotlar",
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/admin/products/{id}": {
      "get": {
        "operationId": "ProductsAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mahsulot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductAdminResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bitta mahsulot",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "description": "Faqat yuborilgan maydonlar o‘zgaradi. `slug` o‘zgarmaydi.\n\n⚠ `sqmPerPallet` / `weightPerPallet` o‘zgarishi ESKI buyurtmalarga ta’sir qilmaydi — ular buyurtma paytidagi qiymatni saqlaydi.",
        "operationId": "ProductsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangilangan mahsulot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductAdminResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Maydonlar noto‘g‘ri, zavod yoki o‘lcham topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mahsulotni tahrirlash",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "description": "Yozuv bazadan o‘chirilmaydi — `isActive: false`, ochiq katalogdan yo‘qoladi. Buyurtma tarixi saqlanib qolishi uchun.\n\nQaytarish: `PATCH { \"isActive\": true }`.",
        "operationId": "ProductsAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "O‘chirilgan mahsulot (isActive: false)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProductAdminResponseDto"
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mahsulotni o‘chirish (soft delete)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/products/{id}/similar": {
      "get": {
        "description": "Faqat admin bog‘laganlari (avtomatik tanlov bu yerda yo‘q). O‘chirilgan mahsulotlar ham ko‘rinadi — ochiq sahifada ular chiqmaydi.",
        "operationId": "ProductsAdminController_findSimilar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bog‘lanishlar tartib bilan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SimilarProductLinkDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Qo‘lda bog‘langan o‘xshash mahsulotlar",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "description": "Ro‘yxat TO‘LIQ almashtiriladi (qo‘shilmaydi). Bo‘sh massiv — hammasini olib tashlash. Tartib — massiv tartibi.\n\nBog‘lanish bir tomonlama: A → B qo‘yilsa, B sahifasida A o‘zi chiqmaydi.",
        "operationId": "ProductsAdminController_setSimilar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Mahsulot ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSimilarProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yangi bog‘lanishlar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SimilarProductLinkDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Qo‘shimcha ma’lumot (masalan sahifalash)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "O‘ziga bog‘lash, topilmagan ID yoki dublikat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Bu amal uchun rol yetarli emas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Mahsulot topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "O‘xshash mahsulotlarni belgilash",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vodiy Kafel API",
    "description": "Vodiy Kafel savdo platformasining backend API'si.\n\n### Javob formati\nBarcha muvaffaqiyatli javoblar `{ \"data\": ... }` ichida keladi.\nSahifalangan ro'yxatlarda qo'shimcha `meta` bo'ladi.\n\nBarcha xatolar bir xil formatda — `ApiErrorDto` modeliga qarang.\n\n### Autentifikatsiya\nToken `Authorization: Bearer <token>` sarlavhasida yuboriladi.\nYuqoridagi **Authorize** tugmasi orqali kiritishingiz mumkin.\n\n### Muhim qoidalar\n- Mahsulot javoblarida ombordagi **aniq son yo'q** — faqat `stockStatus`.\n- Buyurtma summasini backend o'zi hisoblaydi; frontend yuborgan narx e'tiborga olinmaydi.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status kodi"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Status nomi"
          },
          "message": {
            "description": "Xato matni. Validatsiya xatosida — matnlar massivi, aks holda bitta matn.",
            "oneOf": [
              {
                "type": "string",
                "example": "Mahsulot topilmadi"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "pallets must not be less than 1"
                ]
              }
            ]
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products",
            "description": "So‘rov manzili"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-08T17:02:11.419Z",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "example": "1339b1a6-5cf7-4d0f-9d4b-10bfbd295f49",
            "description": "Kuzatuv ID. Xato haqida xabar berganda shu ID ni ayting — log‘dan aniq so‘rov topiladi."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "ProductListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "name": {
            "type": "string",
            "example": "Lyuks Keramogranit"
          },
          "slug": {
            "type": "string",
            "description": "Katalog manzili uchun — `/products/:slug`",
            "example": "lyuks-keramogranit-60x60"
          },
          "factory": {
            "$ref": "#/components/schemas/ProductFactoryRefDto"
          },
          "size": {
            "$ref": "#/components/schemas/ProductSizeRefDto"
          },
          "surface": {
            "type": "string",
            "enum": [
              "POL",
              "DEVOR"
            ],
            "example": "POL"
          },
          "color": {
            "type": "object",
            "example": "Bej",
            "nullable": true
          },
          "sqmPerPallet": {
            "type": "string",
            "description": "1 paddondagi m². O‘nlik son SATR ko‘rinishida keladi — u pul hisobiga kiradi va `float` ga aylantirilsa aniqlik yo‘qolardi (CLAUDE.md qoida 7).",
            "example": "1.44"
          },
          "weightPerPallet": {
            "type": "string",
            "description": "1 paddon og‘irligi (kg), satr ko‘rinishida",
            "example": "1250.5"
          },
          "primaryImageUrl": {
            "type": "object",
            "description": "Kartada ko‘rsatiladigan birinchi surat. Butun media ro‘yxati `GET /products/:slug` da keladi.",
            "example": "/uploads/products/lyuks-1.jpg",
            "nullable": true
          },
          "availability": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "description": "🔒 FAQAT ikki holat. Ombordagi aniq son hech qachon berilmaydi (TZ 3.2). Uch rangli indikator (🟢🟡🔴) auth bor joyda — optom mijoz kabinetida va admin panelida.",
            "example": "AVAILABLE"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "factory",
          "size",
          "surface",
          "sqmPerPallet",
          "weightPerPallet",
          "availability"
        ]
      },
      "PaginatedProductListItemResponseDtoDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductListItemResponseDto"
            }
          },
          "total": {
            "type": "number",
            "example": 128,
            "description": "Jami elementlar soni"
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 7,
            "description": "Jami sahifalar soni"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "ProductDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "name": {
            "type": "string",
            "example": "Lyuks Keramogranit"
          },
          "slug": {
            "type": "string",
            "description": "Katalog manzili uchun — `/products/:slug`",
            "example": "lyuks-keramogranit-60x60"
          },
          "factory": {
            "$ref": "#/components/schemas/ProductFactoryRefDto"
          },
          "size": {
            "$ref": "#/components/schemas/ProductSizeRefDto"
          },
          "surface": {
            "type": "string",
            "enum": [
              "POL",
              "DEVOR"
            ],
            "example": "POL"
          },
          "color": {
            "type": "object",
            "example": "Bej",
            "nullable": true
          },
          "sqmPerPallet": {
            "type": "string",
            "description": "1 paddondagi m². O‘nlik son SATR ko‘rinishida keladi — u pul hisobiga kiradi va `float` ga aylantirilsa aniqlik yo‘qolardi (CLAUDE.md qoida 7).",
            "example": "1.44"
          },
          "weightPerPallet": {
            "type": "string",
            "description": "1 paddon og‘irligi (kg), satr ko‘rinishida",
            "example": "1250.5"
          },
          "primaryImageUrl": {
            "type": "object",
            "description": "Kartada ko‘rsatiladigan birinchi surat. Butun media ro‘yxati `GET /products/:slug` da keladi.",
            "example": "/uploads/products/lyuks-1.jpg",
            "nullable": true
          },
          "availability": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "description": "🔒 FAQAT ikki holat. Ombordagi aniq son hech qachon berilmaydi (TZ 3.2). Uch rangli indikator (🟢🟡🔴) auth bor joyda — optom mijoz kabinetida va admin panelida.",
            "example": "AVAILABLE"
          },
          "description": {
            "type": "object",
            "example": "Yuqori sifatli keramogranit, sirti silliq.",
            "nullable": true
          },
          "media": {
            "description": "Suratlar va 360° materiallar — ko‘rsatish tartibida.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductMediaResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "factory",
          "size",
          "surface",
          "sqmPerPallet",
          "weightPerPallet",
          "availability",
          "media"
        ]
      },
...
      "ProductAdminResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmtz0a1b2c3d4e5f6g7h8i9j"
          },
          "name": {
            "type": "string",
            "example": "Lyuks Keramogranit"
          },
          "slug": {
            "type": "string",
            "example": "lyuks-keramogranit-60x60"
          },
          "factory": {
            "$ref": "#/components/schemas/ProductFactoryRefDto"
          },
          "size": {
            "$ref": "#/components/schemas/ProductSizeRefDto"
          },
          "surface": {
            "type": "string",
            "enum": [
              "POL",
              "DEVOR"
            ],
            "example": "POL"
          },
          "color": {
            "type": "string",
            "example": "Bej",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "sqmPerPallet": {
            "type": "string",
            "example": "1.44"
          },
          "weightPerPallet": {
            "type": "string",
            "example": "32.5"
          },
          "viewCount": {
            "type": "number",
            "example": 0
          },
          "isActive": {
            "type": "boolean",
            "description": "`false` — o‘chirilgan (soft delete), ochiq katalogda yo‘q",
            "example": true
          },
          "stock": {
            "$ref": "#/components/schemas/ProductStockSummaryDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-13T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-13T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "factory",
          "size",
          "surface",
          "sqmPerPallet",
          "weightPerPallet",
          "viewCount",
          "isActive",
          "stock",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Lyuks Keramogranit",
            "maxLength": 150
          },
          "factoryId": {
            "type": "string",
            "description": "Zavod ID"
          },
          "sizeId": {
            "type": "string",
            "description": "O‘lcham ID"
          },
          "surface": {
            "type": "string",
            "enum": [
              "POL",
              "DEVOR"
            ],
            "example": "POL"
          },
          "color": {
            "type": "string",
            "example": "Bej",
            "maxLength": 60
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "sqmPerPallet": {
            "type": "string",
            "description": "1 paddondagi m² — kalkulyator uchun MAJBURIY. Satr ko‘rinishida (float xatosi bo‘lmasligi uchun), verguldan keyin 4 tagacha raqam.",
            "example": "1.44"
          },
          "weightPerPallet": {
            "type": "string",
            "description": "1 paddon og‘irligi (kg) — transport soni shu asosda hisoblanadi. Satr ko‘rinishida, verguldan keyin 3 tagacha raqam.",
            "example": "1250.5"
          },
          "isActive": {
            "type": "boolean",
            "description": "O‘chirilgan mahsulotni qaytarish: `true`. `DELETE` ham shu maydonni `false` qiladi."
          }
        }
      },
...
    }
  }
}